+++ rust/bridge/shared/testing/src/net.rs
@@ -11,7 +11,7 @@ use libsignal_bridge_types::net::{ConnectionManager, TokioAsyncContext};
 use libsignal_core::E164;
 use libsignal_net::cdsi::{CdsiProtocolError, LookupError, LookupResponse, LookupResponseEntry};
 use libsignal_net::infra::errors::RetryLater;
-use libsignal_net::infra::ws2::attested::AttestedProtocolError;
+use libsignal_net::infra::ws::attested::AttestedProtocolError;
 use libsignal_protocol::{Aci, Pni};
 use nonzero_ext::nonzero;
 use uuid::Uuid;